Add convenience UIPickerView rx extensions example
