+++ RxExample/RxExample/Examples/UIPickerViewExample/SimplePickerViewExampleViewController.swift
@@ -21,111 +21,28 @@ final class SimplePickerViewExampleViewController: ViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        Observable.just([[1, 2, 3], [5, 8, 13], [21, 34]])
-            .bind(to: pickerView1.rx.items(adapter: CustomStringPickerViewAdapter()))
+        Observable.just([1, 2, 3])
+            .bind(to: pickerView1.rx.itemTitles) { _, item in
+                return "\(item)"
+            }
             .disposed(by: disposeBag)
         
-        Observable.just([[1, 2, 3], [5, 8, 13], [21, 34]])
-            .bind(to: pickerView2.rx.items(adapter: CustomAttributedStringPickerViewAdapter()))
+        Observable.just([1, 2, 3])
+            .bind(to: pickerView2.rx.itemAttributedTitles) { _, item in
+                return NSAttributedString(string: "\(item)",
+                                          attributes: [
+                                            NSForegroundColorAttributeName: UIColor.cyan,
+                                            NSUnderlineStyleAttributeName: NSUnderlineStyle.styleDouble.rawValue
+                                        ])
+            }
             .disposed(by: disposeBag)
         
-        Observable.just([[1, 2, 3], [5, 8, 13], [21, 34]])
-            .bind(to: pickerView3.rx.items(adapter: PickerViewViewAdapter()))
+        Observable.just([UIColor.red, UIColor.green, UIColor.blue])
+            .bind(to: pickerView3.rx.items) { _, item, _ in
+                let view = UIView()
+                view.backgroundColor = item
+                return view
+            }
             .disposed(by: disposeBag)
     }
 }
-
-final class CustomStringPickerViewAdapter
-    : NSObject
-    , UIPickerViewDataSource
-    , UIPickerViewDelegate
-    , RxPickerViewDataSourceType {
-    typealias Element = [[CustomStringConvertible]]
-    private var items: [[CustomStringConvertible]] = []
-    
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return items.count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return items[component].count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
-        return items[component][row].description
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, observedEvent: Event<Element>) {
-        UIBindingObserver(UIElement: self) { (adapter, items) in
-            adapter.items = items
-            pickerView.reloadAllComponents()
-        }.on(observedEvent)
-    }
-}
-
-final class CustomAttributedStringPickerViewAdapter
-    : NSObject
-    , UIPickerViewDataSource
-    , UIPickerViewDelegate
-    , RxPickerViewDataSourceType {
-    typealias Element = [[CustomStringConvertible]]
-    private var items: [[CustomStringConvertible]] = []
-    
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return items.count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return items[component].count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, attributedTitleForRow row: Int, forComponent component: Int) -> NSAttributedString? {
-        return NSAttributedString(string: items[component][row].description,
-                                  attributes: [
-                                    NSForegroundColorAttributeName: UIColor.cyan,
-                                    NSUnderlineStyleAttributeName: NSUnderlineStyle.styleDouble.rawValue
-                                            ])
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, observedEvent: Event<Element>) {
-        UIBindingObserver(UIElement: self) { (adapter, items) in
-            adapter.items = items
-            pickerView.reloadAllComponents()
-            }.on(observedEvent)
-    }
-}
-
-final class PickerViewViewAdapter
-    : NSObject
-    , UIPickerViewDataSource
-    , UIPickerViewDelegate
-    , RxPickerViewDataSourceType {
-    typealias Element = [[CustomStringConvertible]]
-    private var items: [[CustomStringConvertible]] = []
-    
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return items.count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return items[component].count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
-        let label = UILabel()
-        label.text = items[component][row].description
-        label.textColor = UIColor.orange
-        label.font = UIFont.preferredFont(forTextStyle: .headline)
-        label.textAlignment = .center
-        return label
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, observedEvent: Event<Element>) {
-        UIBindingObserver(UIElement: self) { (adapter, items) in
-            adapter.items = items
-            pickerView.reloadAllComponents()
-            }.on(observedEvent)
-    }
-}
-
-
